InstrStopNode align with gp7

+++ src/include/executor/instrument.h
@@ -125,8 +125,8 @@ extern Instrumentation *InstrAlloc(int n, int instrument_options,
 								   bool async_mode);
 extern void InstrInit(Instrumentation *instr, int instrument_options);
 extern void InstrStartNode(Instrumentation *instr);
-extern void InstrStopNode(Instrumentation *instr, double nTuples);
-extern void InstrUpdateTupleCount(Instrumentation *instr, double nTuples);
+extern void InstrStopNode(Instrumentation *instr, uint64 nTuples);
+extern void InstrUpdateTupleCount(Instrumentation *instr, uint64 nTuples);
 
 extern void InstrEndLoop(Instrumentation *instr);
 extern void InstrAggNode(Instrumentation *dst, Instrumentation *add);
